Fix wrong linking to duplicated headings
When there are more than 1 headings with the same name, they currently
end up having the same link name which breaks the "href" functionality
(not navigated to the heading after pointer click).

This commit fixes the issue by introducing associative array that tracks
amount of duplicated headings and appends suffix "-<number>" to the
link.

Signed-off-by: Ondrej Vasko <ondrej.vaskoo@gmail.com>